// src/Features/Common/Enums/status.enum.ts
export enum orderStatus {
  ORDERPLACED = "order placed",
  SHIPPED = "shipped",
  OUTFORDELIVERY = "out for delivery",
  DELIVERED = "delivered",
  DELAYED = "delayed",
  FAILED = "failed",
}
export enum paymentStatus {
  PENDING = "pending",
  SUCCESS = "success",
  FAILED = "failed",
  CANCELLED = "cancelled",
  REFUNDED = "refunded",
}
export enum paymentMode {
  COD = "cod",
  ONLINE = "online",
}